+++ server/routes/_version.get.ts
@@ -1,0 +1,3 @@
+export default defineEventHandler(() => ({
+  version: useRuntimeConfig().public.version,
+}));